Fix for tatsunoko vs capcom and those wii games getting corrupted memory error + a couple of (serious...) fixes for some of my mistakes, and some warning fixes
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@3411 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Common/Src/FileUtil.cpp
@@ -633,7 +633,7 @@ bool ReadFileToString(bool text_file, const char *filename, std::string &str)
 	if (!f)
 		return false;
 	fseek(f, 0, SEEK_END);
-	size_t len = ftell(f);
+	size_t len = (size_t)ftell(f);
 	fseek(f, 0, SEEK_SET);
 	char *buf = new char[len + 1];
 	buf[fread(buf, 1, len, f)] = 0;
